ManifoldGroupTesting.jl

ManifoldGroupTesting.jl contains several useful functions to test new group implementations.

ManifoldGroupTesting.check_action_morphism — Method

check_action_morphism(α::GroupAction, χ1, χ2, p)

For a group action α,

\[α(χ_1, α(χ_2, p)) = α(m(χ_1, χ_2), p)\]

where for a left action

\[m(χ_1, χ_2) = χ_1 χ_2\]

and for a right action

\[m(χ_1, χ_2) = χ_2 χ_1\]

ManifoldGroupTesting.check_adjoint_action — Method

check_adjoint_action(G, grp_rep, alg_rep, χ, ξ)

The group representation $ρ$ and algebra representation $ρ$ commute with the adjoint action:

\[ρ(χ) ρ(ξ) ρ(χ)^{-1} = ρ(Ad_χ (ξ))\]

ManifoldGroupTesting.check_adjoint_action_in_alg — Method

check_adjoint_action_in_alg(G, χ, ξ)

The adjoint action of χ on ξ is an element of Alg(G):

\[Ad_{χ}ξ ∈ Alg(G)\]

ManifoldGroupTesting.check_alg_rep — Method

check_alg_rep(G, alg_rep, ξ1, ξ2)

The algebra representation $ρ$ is an algebra morphism.

\[ρ([ξ_1, ξ_2]) = [ρ(ξ_1), ρ(ξ_2)]\]

where the latter is a matrix commutator.

ManifoldGroupTesting.check_apply_diff_group — Method

check_apply_diff_group(
    A::AbstractGroupAction{TAD}, # group action G ⊂ Diff(M)
    χ, # element of G
    ξ, # element of Alg(G)
    p, # element of M
)

This should hold for any group action $A$ on any manifold. If you define $π_p(χ) := A(χ, p)$ for $χ ∈ G$ and $p ∈ M$, and define, for $ξ ∈ Alg(G)$, $T_R(χ, ξ) := ξχ$ (the right translation), and $T_L(χ, ξ) := χξ$ (the left translation), then we have the identity:

\[⟨Dπ_{p}(χ), T(χ, ξ)⟩ = ⟨Dπ_{A(χ,p)}(1), ξ⟩\]

where, for a left action, $T$ is the right translation, and for a right action, $T$ is the left translation.

ManifoldGroupTesting.check_apply_diff_group_at_id — Function

check_apply_diff_group_at_id(G, side::GroupActionSide)

The left group operation action on itself $α(χ_1)χ_2$ is either (left side)

\[α(χ_1)χ_2 = χ_1 χ_2\]

or (right side)

\[α(χ_1)χ_2 = χ_2 χ_1^{-1}\]

Now fix $χ_2 = 1$ (1 is the identity of $G$) and define $f : G → G$ by $f(χ) := α(χ) 1$. Since $f(1) = 1$, its differential at identity is a map $Alg(G) → Alg(G)$. This map is either

$Id$ (left side)

$-Id$ (right side)

ManifoldGroupTesting.check_apply_morphism_Identity — Method

check_apply_morphism_Identity(α::AbstractGroupAction, p)

For any action $α$ and $p$ in the manifold, one has

\[α(1, p) = p\]

where $1$ denotes $Identity(G)$ where $G$ is the action group.

ManifoldGroupTesting.check_exp_ad — Method

$Ad$ and $exp$ commute:

\[Ad_{exp(ξ_1)}ξ_2 = exp(ad_{ξ_1}) ξ_2\]

ManifoldGroupTesting.check_exp_lie_point — Method

check_exp_lie_point(G, ξ)

The Lie group exponential sends the vector ξ to an element in the group.

ManifoldGroupTesting.check_grp_rep_Identity — Method

check_grp_rep_Identity(G)

The representation works at the Identity(G) point.

ManifoldGroupTesting.check_grp_rep_compose — Method

check_grp_rep_compose(G, ρ, χ1, χ2)

The group representation $ρ$ commutes with composition.

\[ ρ(χ_1 χ_2) = ρ(χ_1) ρ(χ_2)\]

where the latter is a matrix multiplication.

ManifoldGroupTesting.check_inv_diff — Method

check_inv_diff(
  G, # Group
  χ, # group element
  ξ, # Lie algebra element
  side::Manifolds.GroupActionSide,
  )

Test the differential of the inverse on a Lie group G. Denote this inverse by $I(χ) := χ^{-1}$. If the left and right transports are $T_L(χ,ξ) := χξ$ and $T_R(χ,ξ) := ξχ$ respectively, then

\[⟨DI(χ), T_L(χ,ξ)⟩ = -T_R(χ^{-1}, ξ)\]

and

\[⟨DI(χ), T_R(χ,ξ)⟩ = -T_L(χ^{-1}, ξ)\]

ManifoldGroupTesting.check_inv_rep — Method

check_inv_rep(G, grp_rep, χ)

The group representation $ρ$ commutes with the inverse.

\[ρ(χ)^{-1} = ρ(χ^{-1})\]

ManifoldGroupTesting.check_switch_action_direction — Method

check_switch_action_direction(α::GroupAction, χ, p)

For any group action $α$ and its dual $α'$ obtained by switch_direction one has

\[α(χ,p) = α'(χ^{-1}, p)\]

ManifoldGroupTesting.check_trivial_infinitesimal_action — Function

check_trivial_infinitesimal_action(A::GroupAction, p)

For an action $α$ and a point $p∈ M$, consider the function $f : G → M$ defined as $f(χ) := α(χ, p)$. The tangent map at identity $D := T f(1)$ is a linear map and sends zero to zero:

\[D 0 = 0\]

ManifoldGroupTesting.translate_from_id — Method

The left translation $T_L(g,ξ) = gξ$.

ManifoldGroupTesting.translate_from_id — Method

The right translation $T_R(g,ξ) = ξg$.